fix: Compiled code cannot source the python code by path
